임시 커밋

+++ src/main/java/com/yjh/whoplayer/controller/data/UserController.java
@@ -21,7 +21,7 @@ public class UserController {
     }
 
     @PostMapping("/update-game")
-    public void updateGame(@RequestBody Integer erUid) {
+    public void updateGame(@RequestBody Long erUid) {
         gameService.updateGame(erUid);
     }
 }
